LPS-174373 extract code from AutoDeployListener to a more appropriate classes
