+++ modules/apps/batch-engine/batch-engine-service/src/main/java/com/liferay/batch/engine/internal/unit/BatchEngineUnitConfigurationHelper.java
@@ -1,0 +1,83 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.batch.engine.internal.unit;
+
+import com.liferay.batch.engine.unit.BatchEngineUnitConfiguration;
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.model.Company;
+import com.liferay.portal.kernel.service.CompanyLocalService;
+import com.liferay.portal.kernel.service.UserLocalService;
+import com.liferay.portal.kernel.util.PropsKeys;
+import com.liferay.portal.kernel.util.PropsUtil;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Matija Petanjek
+ */
+@Component(service = BatchEngineUnitConfigurationHelper.class)
+public class BatchEngineUnitConfigurationHelper {
+
+	public BatchEngineUnitConfiguration updateBatchEngineUnitConfiguration(
+		BatchEngineUnitConfiguration batchEngineUnitConfiguration) {
+
+		if (batchEngineUnitConfiguration.companyId == 0) {
+			if (_log.isWarnEnabled()) {
+				_log.warn("Using default company ID for this batch process");
+			}
+
+			try {
+				Company company = _companyLocalService.getCompanyByWebId(
+					PropsUtil.get(PropsKeys.COMPANY_DEFAULT_WEB_ID));
+
+				batchEngineUnitConfiguration.companyId = company.getCompanyId();
+			}
+			catch (PortalException portalException) {
+				_log.error("Unable to get default company ID", portalException);
+			}
+		}
+
+		if (batchEngineUnitConfiguration.userId == 0) {
+			if (_log.isWarnEnabled()) {
+				_log.warn("Using default user ID for this batch process");
+			}
+
+			try {
+				batchEngineUnitConfiguration.userId =
+					_userLocalService.getUserIdByScreenName(
+						batchEngineUnitConfiguration.companyId,
+						PropsUtil.get(PropsKeys.DEFAULT_ADMIN_SCREEN_NAME));
+			}
+			catch (PortalException portalException) {
+				_log.error("Unable to get default user ID", portalException);
+			}
+		}
+
+		return batchEngineUnitConfiguration;
+	}
+
+	private static final Log _log = LogFactoryUtil.getLog(
+		BatchEngineUnitConfigurationHelper.class);
+
+	@Reference
+	private CompanyLocalService _companyLocalService;
+
+	@Reference
+	private UserLocalService _userLocalService;
+
+}
