+++ modules/apps/batch-engine/batch-engine-service/src/main/java/com/liferay/batch/engine/internal/unit/BatchEngineUnitProcessorImpl.java
@@ -1,0 +1,172 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.batch.engine.internal.unit;
+
+import com.liferay.batch.engine.BatchEngineImportTaskExecutor;
+import com.liferay.batch.engine.BatchEngineTaskExecuteStatus;
+import com.liferay.batch.engine.BatchEngineTaskOperation;
+import com.liferay.batch.engine.constants.BatchEngineImportTaskConstants;
+import com.liferay.batch.engine.internal.auto.deploy.BatchEngineAutoDeployListener;
+import com.liferay.batch.engine.model.BatchEngineImportTask;
+import com.liferay.batch.engine.service.BatchEngineImportTaskLocalService;
+import com.liferay.batch.engine.unit.BatchEngineUnit;
+import com.liferay.batch.engine.unit.BatchEngineUnitConfiguration;
+import com.liferay.batch.engine.unit.BatchEngineUnitProcessor;
+import com.liferay.petra.executor.PortalExecutorManager;
+import com.liferay.petra.io.StreamUtil;
+import com.liferay.petra.io.unsync.UnsyncByteArrayOutputStream;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.service.CompanyLocalService;
+import com.liferay.portal.kernel.service.UserLocalService;
+import com.liferay.portal.kernel.util.File;
+import com.liferay.portal.kernel.util.StringUtil;
+import com.liferay.portal.kernel.util.Validator;
+
+import java.io.InputStream;
+
+import java.util.concurrent.ExecutorService;
+import java.util.zip.ZipEntry;
+import java.util.zip.ZipOutputStream;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Matija Petanjek
+ */
+@Component(service = BatchEngineUnitProcessor.class)
+public class BatchEngineUnitProcessorImpl implements BatchEngineUnitProcessor {
+
+	@Override
+	public void processBatchEngineUnits(
+		Iterable<BatchEngineUnit> batchEngineUnits) {
+
+		for (BatchEngineUnit batchEngineUnit : batchEngineUnits) {
+			try {
+				_processBatchEngineUnit(batchEngineUnit);
+
+				if (_log.isInfoEnabled()) {
+					_log.info(
+						"Successfully enqueued batch file " +
+							batchEngineUnit.getFileName());
+				}
+			}
+			catch (Exception exception) {
+				if (_log.isDebugEnabled()) {
+					_log.debug(
+						"Ignoring invalid batch file " +
+							batchEngineUnit.getFileName(),
+						exception);
+				}
+			}
+		}
+	}
+
+	private void _processBatchEngineUnit(BatchEngineUnit batchEngineUnit)
+		throws Exception {
+
+		BatchEngineUnitConfiguration batchEngineUnitConfiguration = null;
+		byte[] content = null;
+		String contentType = null;
+
+		if (batchEngineUnit.isValid()) {
+			batchEngineUnitConfiguration =
+				_batchEngineUnitConfigurationHelper.
+					updateBatchEngineUnitConfiguration(
+						batchEngineUnit.getBatchEngineUnitConfiguration());
+
+			UnsyncByteArrayOutputStream compressedUnsyncByteArrayOutputStream =
+				new UnsyncByteArrayOutputStream();
+
+			try (InputStream inputStream = batchEngineUnit.getDataInputStream();
+				ZipOutputStream zipOutputStream = new ZipOutputStream(
+					compressedUnsyncByteArrayOutputStream)) {
+
+				zipOutputStream.putNextEntry(
+					new ZipEntry(batchEngineUnit.getDataFileName()));
+
+				StreamUtil.transfer(inputStream, zipOutputStream, false);
+			}
+
+			content = compressedUnsyncByteArrayOutputStream.toByteArray();
+
+			contentType = _file.getExtension(batchEngineUnit.getDataFileName());
+		}
+
+		if ((batchEngineUnitConfiguration == null) || (content == null) ||
+			Validator.isNull(contentType)) {
+
+			throw new IllegalStateException(
+				"Invalid batch engine file " + batchEngineUnit.getFileName());
+		}
+
+		ExecutorService executorService =
+			_portalExecutorManager.getPortalExecutor(
+				BatchEngineAutoDeployListener.class.getName());
+
+		BatchEngineImportTask batchEngineImportTask =
+			_batchEngineImportTaskLocalService.addBatchEngineImportTask(
+				null, batchEngineUnitConfiguration.companyId,
+				batchEngineUnitConfiguration.userId, 100,
+				batchEngineUnitConfiguration.callbackURL,
+				batchEngineUnitConfiguration.className, content,
+				StringUtil.toUpperCase(contentType),
+				BatchEngineTaskExecuteStatus.INITIAL.name(),
+				batchEngineUnitConfiguration.fieldNameMappingMap,
+				BatchEngineImportTaskConstants.IMPORT_STRATEGY_ON_ERROR_FAIL,
+				BatchEngineTaskOperation.CREATE.name(),
+				batchEngineUnitConfiguration.parameters,
+				batchEngineUnitConfiguration.taskItemDelegateName);
+
+		executorService.submit(
+			() -> {
+				_batchEngineImportTaskExecutor.execute(batchEngineImportTask);
+
+				if (_log.isInfoEnabled()) {
+					_log.info(
+						"Successfully deployed batch engine file " +
+							batchEngineUnit.getFileName());
+				}
+			});
+	}
+
+	private static final Log _log = LogFactoryUtil.getLog(
+		BatchEngineUnitProcessorImpl.class);
+
+	@Reference
+	private BatchEngineImportTaskExecutor _batchEngineImportTaskExecutor;
+
+	@Reference
+	private BatchEngineImportTaskLocalService
+		_batchEngineImportTaskLocalService;
+
+	@Reference
+	private BatchEngineUnitConfigurationHelper
+		_batchEngineUnitConfigurationHelper;
+
+	@Reference
+	private CompanyLocalService _companyLocalService;
+
+	@Reference
+	private File _file;
+
+	@Reference
+	private PortalExecutorManager _portalExecutorManager;
+
+	@Reference
+	private UserLocalService _userLocalService;
+
+}
