+++ modules/apps/batch-engine/batch-engine-service/src/main/java/com/liferay/batch/engine/internal/writer/JSONTBatchEngineExportTaskItemWriterImpl.java
@@ -48,7 +48,7 @@ public class JSONTBatchEngineExportTaskItemWriterImpl
 		_unsyncPrintWriter.write(
 			"{\"actions\":\n{\"createBatch\": {\"href\": \"");
 		_unsyncPrintWriter.write("/o/headless-batch-engine/v1.0/import-task/");
-		_unsyncPrintWriter.write(batchEngineImportConfiguration.getClassName());
+		_unsyncPrintWriter.write(batchEngineUnitConfiguration.getClassName());
 		_unsyncPrintWriter.write(
 			"\", \"method\": \"POST\"}},\n\"configuration\":\n");
 		_unsyncPrintWriter.write(
